pyproject.toml
[tool.poetry]
name = "ufcz"
version = "0.1.0"
description = ""
authors = ["jkrbec <jaroslavkrbec@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
